Finished html hard coded restaurant web package
I setup webpack with the correct configurations,
made a webpack.config.js, and re-installed
webpack on this repo with npm (which
automatically made a node_modules folder and
package.json files to track dependencies).
Also made a .gitignore to ignore the node_modules
list because when a user gets the .json files
and does npm install, it will just install
all of the required dependencies anyways, and
dist folder since these are GENERATED bundled
source code, so we dont need to track it since
we can re-bundle/remake the code with npx
webpack from the source files.